test: proposal of autogeneration of
Proposal for how to auto generate the events & config files to be
used in the conformance tests.

When running conformance tests, we need the following:
- config.json
- events.json

Both of which can be automatically generated. This commit makes that
a possibility.

Since this commit is only a proposal, this has only been implemented for
vrrp/vrrpv3-topo-1-1.

Inside our setup directory, we have an empty env.txt file, we should
populate this file with the path to our holo project directory...like
so:

```
SNAPSHOT_DIR=/my/holo/directory
```

When we run the munet topology using `munet -c
vrrp/vrrpv3-topo-1-1/munet.yaml`, there will automatically be
config.json and events.json files for each of the devices in the topo
that are generated.

Work in progress to get the output directory files(ibus.jsonl,
protocol.jsonl, etc...) also included.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>